% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_merch_lookup.R
\name{create_merch_lookup}
\alias{create_merch_lookup}
\title{Create ABS Lookup table for merchandise exports data}
\usage{
create_merch_lookup(path = tempdir())
}
\arguments{
\item{path}{Directory in which downloaded XML should be stored}
}
\value{
A list, each element of which is a \code{tbl_df}
}
\description{
The ABS merchandise exports data
(\url{https://stat.data.abs.gov.au/index.aspx?DatasetCode=MERCH_EXP}) is
supplied with short codes for various data fields, such as industry and
SITC code. This function creates a series of lookup tables that include
English descriptions of these fields.
}
\examples{
create_merch_lookup()
}
