% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_table.R
\name{make_table_launchpad}
\alias{make_table_launchpad}
\title{Create a table for the dashboard or a briefing document}
\usage{
make_table_launchpad(
  data,
  destination = Sys.getenv("R_DJPRLABOURDASH_TABLEDEST", unset = "dashboard"),
  notes = NULL,
  title = "",
  header_row = c("", "Current figure & share (\%)", "Change in latest period",
    "Change in three months", "Change in past year")
)
}
\arguments{
\item{data}{A data frame containing data to summarise}

\item{destination}{"dashboard" or "briefing"}

\item{notes}{Optional notes to add to caption. Source will be inferred
automatically based on the data using \code{caption_auto()}.}

\item{title}{Character vector to use as the table title. Will only be used
when \code{destination} is "briefing".}

\item{header_row}{Top row labels}
}
\description{
Create a table for the dashboard or a briefing document
}
\examples{
# dash_data <- load_dash_data()
\dontrun{
make_table(
  data = filter_dash_data(series_ids = c(
    "A84423354L",
    "A84423242V",
    "A84423466F"
  )),
  row_order = c(
    "A84423354L",
    "A84423242V",
    "A84423466F"
  ),
  highlight_rows = c("A84426256L")
)
}
}
