% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bop.R
\name{read_bop}
\alias{read_bop}
\title{Download, tidy and import ABS balance of payments data}
\usage{
read_bop(path = tempdir())
}
\arguments{
\item{path}{Path where Excel files from the ABS should be stored}
}
\value{
A tibble containing both imports (debits) and credits (exports)
data
}
\description{
Downloads ABS Balance of Payments (5302.0) Tables 21 and 22, goods and
services credits and debits by state.
}
\examples{
\dontrun{
read_bop()
}
}
