% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_services.R
\name{read_services}
\alias{read_services}
\title{Read service trade data}
\usage{
read_services(...)
}
\arguments{
\item{...}{unused}
}
\value{
data.frame
}
\description{
Downloads, parses and cleans the latest service trade data
}
