% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_supp.R
\name{read_supp}
\alias{read_supp}
\title{Download, tidy and import ABS International Trade Supplementary Information}
\usage{
read_supp(
  format = c("cy", "fy"),
  table_no = c(1, 2, 3, 4, 5, 6, 7, 8),
  list = FALSE,
  path = tempdir()
)
}
\arguments{
\item{format}{Selects whether calendar year or financial year data is imported}

\item{table_no}{Selects which tables from the ABS will be imported}

\item{list}{Where multiple tables are requested, if list is TRUE, table outputs exported separately in a list, if FALSE tables are exported in a combined tibble.}

\item{path}{Path where Excel files downloaded from the ABS should be stored}
}
\value{
A tibble containing the relevant table, or a list of tibbles if more than one ABS table has been requested
}
\description{
Downloads Tables 1 through 8, inculsive, of the ABS International Trade: Supplementary Information data.
}
\examples{
read_supp()
}
