% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/djprtradedash-package.R
\docType{package}
\name{djprtradedash-package}
\alias{djprtradedash}
\alias{djprtradedash-package}
\title{djprtradedash: DJPR Trade Dashboard}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Matt Cowgill \email{mattcowgill@gmail.com} (\href{https://orcid.org/0000-0003-0422-3300}{ORCID})

Other contributors:
\itemize{
  \item Department of Jobs, Precincts and Regions (Victoria) [copyright holder, funder]
}

}
\keyword{internal}
