% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_merch.R
\name{read_merch}
\alias{read_merch}
\title{Download merchandise exports data}
\usage{
read_merch(
  path = tempdir(),
  max_date = Sys.Date(),
  min_date = as.Date("2000-01-01"),
  series = c("export", "import")
)
}
\arguments{
\item{path}{Path to directory where CSV files are extracted and stored}

\item{max_date}{The maximum date to include in your data}

\item{min_date}{The minimum date to include in your data}

\item{series}{Selects whether import or export merchandise data is downloaded}
}
\value{
A tibble containing merchandise export data
}
\description{
Obtains merchandise exports data from ABS Data Explorer
(\url{https://explore.data.abs.gov.au/?tm=Merchandise&pg=0}).
}
\details{
The ABS will not supply data frames of over 1m rows using the
ABS.Stat API. For this reason, you cannot download more than 1 year at a
time worth of data using this function, as this is around the point at which
the 1m row limit is reached.
}
\examples{
\dontrun{
read_merch()
}
}
